Add notion of backend pinging
Added new backend callback function 'ping' that used to ensure a
backend is still healthy, based on whatever that backend considers
healthy. The goal is to ensure that unhealthy peers do not remain
leaders.

As an example, consider a backend that is implemented by a separate
Erlang process. The backend could have the ping callback send a ping
message to this process, which then replies with the pong. This
ensures that the separate process is not indefinitely blocked or has
extraordinarily large message queues.

This commit also refactors the leader tick logic, which was becoming
unwieldy. The new ping check would have added yet another nested case
statement. The refactored approach is inspired by the maybe monad in
other languages, but is nonetheless straightforward Erlang code.