Improve stable leadership logic
This commit fixes two issues with the current stable leader logic:
1. Peers incorrectly reporting a stale leader as a valid stable leader.
2. Peers refusing to follow a valid stable leader from a lower epoch.

Recall how stable leadership is handled (prior to this commit).
-- A probe request is sent to other peers asking for their current fact.
-- After receiving a quorum of facts, the latest fact is chosen.
-- If the latest fact has a defined leader, the peer follows that leader.
-- Otherwise, the peer follows the most frequent leader across all facts.
-- If there is no frequent leader, the peer moves on to election.
-- Followers always follow their current leader, ignoring prepare messages
   from other peers.

As implemented, this provides stable leadership in the common case. A new
leader cannot be elected until enough followers abandon the current leader,
and partitioned/restarting peers follow the current stable leader rather
than vying for leadership.

However, there are two problems (mentioned above).

First, it's possible for a new peer to start following a leader just as
followers are about to abandon the leader. These abandoning followers
end up re-following the old leader because the new peer reports it as a
valid stable leader. It is possible for this to happen continuously,
preventing a new leader from ever being elected. Thus, the ensemble
becomes forever unavailable.

Second, its possible for a peer to refuse to follow the current stable
leader, resulting in a peer that remains forever in election. This
occurs when a peer is initially unable to communicate with the current
leader, thus incrementing its epoch to try to elect itself. Since the peer
now has an epoch greater than the epoch of the current leader, the peer
ignores any future prepare requests from the leader after communication
is restored.

This commit makes the following changes to address these issues.

First, peers now remember the epoch/seq at which they abandoned the most
recent leader. When checking for an existing leader, peers now filter out
leaders associated with an epoch/seq earlier then this time.

This change prevents peers from re-following a leader they have already
abandoned, unless the leader has proven it is still valid by incrementing
the sequence counter -- which can only occur if there is a quorum of
followers following that leader.

Second, peers no longer directly move to a new epoch until after a leader
has guaranteed it has a quorum of potential followers. This prevents the
epoch from increasing arbitrarily.

This change is handled by the addition of the prelead and prefollow
states. Upon receiving a prepare message, followers consider the proposed
epoch/leader to be provisional and move into the prefollow state. Once in
the prefollow state, the peer expects to see a new_epoch message with the
same epoch/leader. If a new_epoch message does not arrive within a fixed
time period, the peer gives up on the provisional leader and moves back
into the probe state. If a message does arrive, the peer promotes the
preliminary epoch/leader to the actual epoch/leader and moves into the
following state.

Similarly, after receiving a quorum of prepare responses the leader moves
into the prelead state. The provisional leader then sends the new_epoch
message. If a quorum of replies are received, the leader updates its
epoch to the newly established one and moves into the leading state.

A key component of this design is that peers never increment and then
revert the actual committed epoch. Peers increment a preliminary epoch,
and throw it away if a new epoch is not actually established, but this
preliminary epoch is never committed or used as the actual trusted fact
epoch. The committed epoch still evolves monotonically, with only a
single leader ever leading for a given epoch.

Given this change, partitioned peers no longer increment their epoch when
they cannot talk to the current stable leader. Thus, these peers remain
at the current epoch and will eventually follow the stable leader after
the partition has healed. Thus, addressing the second problem from above.