Change kmodify API to provide future object version
This commit changes riak_ensemble_peer:kmodify to provide the future
object version (the epoch/seq that the given object will be committed
with after modification) to the provided callback function.

This extension makes it possible to write modification functions that
internally version objects, have a need for a monotonic sequence id,
etc.