Fix root ensemble gossip bug
Prior to this commit, the mechanism used by the root leader to
gossip its state to the local ensemble manager was routed through
the riak_ensemble_routers. However, the ensemble routers use the
manager state to route the request to the appropriate ensemble
leader. If the manager was outdated and believed a now partitioned
peer was the leader, the new leader would never be able to properly
route the update request to itself to update the outdated manager.

Since the leader is routing a request to itself, this commit changes
the root leader to directly send the update request to itself by pid,
bypassing the router logic. This approach always works, and fixes the
above issue.