Fix view transition corner case
Prior to this commit, a newly established leader would set view_vsn to
{Epoch, 0}. However, this could lead to the same view_vsn being used
twice for two different views. Specifically, if the leader changes
while there is an outstanding joint view, the new leader could report
the joint view to the ensemble manager with a vsn of {Epoch,0} as well
as report the transitioned view on the next tick with the same vsn.

This commit changes the leader to use {Epoch,-1} for the view_vsn when
newly established, which will always be less than the {Epoch,0} vsn
that could potentially be used to commit the view transition on the
next tick.