Add test for dataloss while trusting pending peers
This test will fail when we automatically trust pending peers.
This allows us to reach a new quorum of peers that don't contain data,
because we aren't syncing when pending peers startup. This can result in
existing data being returned as `notfound`. Fixing this failing test
requires syncing data before transitioning to the final view.

Test for #17