Remove pending state from riak_ensemble_peer
The pending state allows pending peers to act trusted, when they
shouldn't be. It is also unnecessary. Removal of the pending state makes
pending_dataloss_test.erl pass and completes #17.

Pending views are only needed to ensure propogation of views so that
peers may be started dynamically on remote nodes before the leader
transitions to using them. If the leader tried to use the peers before
they were up, he would lose leadership and the peers may never be
started because they require propogation to the manager from the leader
to start them.  Therefore pending views were introduced to help ensure
those peers were started and stable leadership was maintained. However,
once those peers are started they are just normal peers and don't need
to do anything special outside the protocol. Therefore there is no need
for a separate pending state.

A better description by @jtuple is in this comment:
https://github.com/basho/riak_ensemble/commit/01dda5a2d1e0fa56a8096eacc7a1c9495d3bce34#commitcomment-5888125