Make several hardcoded settings user configurable
As a first pass, we directly query configurable settings directly from
the application environment variables. In the future, we should profile
and see if the added contention against the application ETS table matters.
If so, we could consider caching settings using a approach such as the
one used by sidejob_config.