gossip ClusterState to root manager on join/remove
When a node is newly joined, it is likely ensembles will be manipulated
on the newly joined node. However, future ensemble ops will fail until
the newly joined cluster member has been learned by the root manager on
the next tick. These changes ensure that the gossip happens immediately
on join and remove and prevents spurious errors relating to asynch
gossip, that must be retried. It is just an optimization.

Fixes #12 for join/remove case. Other cases were not changed as the
analysis seems less straightforward, and I'm not sure changes are even
required.