Avoid crash on remote manager down
Convert the exit exception from a gen_server call to a remote manager to
an error code. This makes it more natural to handle the situation.
This fix is related to issue basho/riak#559, to prevent crashing when
operating in mixed clusters where ensembles are not enabled on all
nodes.