Fix/improve synctree_pure and corrupt_exchange tests
The recent change to synctree_leveldb to allow sharing LevelDB instances
introduced a necessary ETS table that was not being created in the
synctree_pure test. This commit fixes the test.

While investigating the above failure, it become apparent that the
corrupt_exchange_test was somewhat flaky and would fail on occasion.
This commit improves the test by replacing a timeout with an explicit
test condition.