Add all_or_quorum mode to wait_for_quorum code
This new mode will start out behaving like the 'quorum' mode, but then
if a quorum is achieved, it waits for an extra period of time to collect
additional replies that may or may not come in.

This is useful for optimizing the read-not-found case, because if we get
back replies from all peers in the ensemble, we can verify whether or
not any partial writes exist. This can allow us to skip writing a
tombstone which would otherwise be extraneous.