Add debug_local_get function test purposes
The new function riak_ensemble_peer:debug_local_get can be used to pull
a value from a specific ensemble peer directly, without going through
the normal code path for a get. This is useful for a unit test I want to
write, where we need the ability to check if a specific peer has a
tombstone present for a given key. It may also be useful for other tests
and/or debugging in the future.